4.23*: bloglist expansion, step11 completed

+++ part4/tests/blogs_api.test.js
@@ -20,12 +20,23 @@ const initialBlogs = [
   },
 ];
 
+const initialUser = {
+  username: "Utente1",
+  password: "segreto",
+  name: "Utente_test1",
+};
+
+const userLogin = { username: "Utente1", password: "segreto" };
+
 beforeEach(async () => {
   await Blog.deleteMany({});
   let newBlog = new Blog(initialBlogs[0]);
   await newBlog.save();
   newBlog = new Blog(initialBlogs[1]);
   await newBlog.save();
+
+  await User.deleteMany({});
+  await api.post("/api/users").send(initialUser).set("Accept", "application/json");
 });
 
 test("GET to /api/blogs returns blogs as JSON", async () => {
@@ -41,6 +52,8 @@ test("GET to /api/blogs return correct blogs number", async () => {
 });
 
 test("POST to /api/blogs create a new entry", async () => {
+  const authorization = await api.post("/api/login").send(userLogin).set("Accept", "application/json");
+
   const addBlog = {
     title: "Breaking Good",
     author: "Accendi Unlume",
@@ -50,16 +63,19 @@ test("POST to /api/blogs create a new entry", async () => {
 
   await api
     .post("/api/blogs")
+    .auth(authorization.body.token, { type: "bearer" })
     .send(addBlog)
     .set("Accept", "application/json")
     .expect("Content-Type", /json/)
     .expect(201);
 
   const response = await api.get("/api/blogs");
-  expect(response.body.length).toBe(3);
+  expect(response.body.length).toBe(initialBlogs.length + 1);
 });
 
 test("POST with empty likes returns them as 0 ", async () => {
+  const authorization = await api.post("/api/login").send(userLogin).set("Accept", "application/json");
+
   const addBlogNoLikes = {
     title: "Breaking Good",
     author: "Accendi Unlume",
@@ -68,6 +84,7 @@ test("POST with empty likes returns them as 0 ", async () => {
 
   await api
     .post("/api/blogs")
+    .auth(authorization.body.token, { type: "bearer" })
     .send(addBlogNoLikes)
     .set("Accept", "application/json")
     .expect("Content-Type", /json/)
@@ -78,6 +95,8 @@ test("POST with empty likes returns them as 0 ", async () => {
 });
 
 test("Title and url are required. If missing status is 400", async () => {
+  const authorization = await api.post("/api/login").send(userLogin).set("Accept", "application/json");
+
   const addBlogNoTitle = {
     author: "Accendi Unlume",
     url: "www.ggww.com",
@@ -90,16 +109,37 @@ test("Title and url are required. If missing status is 400", async () => {
     likes: 4,
   };
 
-  await api.post("/api/blogs").send(addBlogNoTitle).set("Accept", "application/json").expect(400);
-  await api.post("/api/blogs").send(addBlogNoUrl).set("Accept", "application/json").expect(400);
+  await api
+    .post("/api/blogs")
+    .auth(authorization.body.token, { type: "bearer" })
+    .send(addBlogNoTitle)
+    .set("Accept", "application/json")
+    .expect(400);
+  await api
+    .post("/api/blogs")
+    .auth(authorization.body.token, { type: "bearer" })
+    .send(addBlogNoUrl)
+    .set("Accept", "application/json")
+    .expect(400);
   const response = await api.get("/api/blogs");
   expect(response.body.length).toBe(2);
 });
 
 test("DELETE to /api/blogs/:id works properly", async () => {
+  const addBlog = {
+    title: "Breaking Good",
+    author: "Accendi Unlume",
+    url: "www.ggww.com",
+    likes: 2,
+  };
+
+  const authorization = await api.post("/api/login").send(userLogin).set("Accept", "application/json");
+
+  await api.post("/api/blogs").auth(authorization.body.token, { type: "bearer" }).send(addBlog);
+
   const response = await api.get("/api/blogs");
-  const idToDelete = response.body[1].id;
-  await api.delete(`/api/blogs/${idToDelete}`).expect(200);
+  const idToDelete = response.body[2].id;
+  await api.delete(`/api/blogs/${idToDelete}`).auth(authorization.body.token, { type: "bearer" }).expect(200);
 });
 
 test("PUT to /api/blogs/:id works properly", async () => {
